requirements.txt
requests
termcolor
python-dotenv
rich